Revert "Initial cleanup"
