--- .gitignore
@@ -1,141 +1,0 @@
-# Temporary editor files
-*~
-.*.swp
-.DS_Store
-
-*.py[cod]
-
-# C extensions
-*.so
-# Try letting str2lc_py.c be in repo to remove dependency on Cython for users to install PINT
-# pint/cutils/str2ld_py.c
-
-# pickle files
-*.pickle.gz
-
-*.npy
-
-# Packages
-*.egg
-*.egg-info
-dist
-build
-eggs
-parts
-bin
-var
-sdist
-develop-eggs
-.installed.cfg
-lib
-lib64
-__pycache__
-.eggs
-
-# Installer logs
-pip-log.txt
-
-# Unit test / coverage reports
-.coverage
-.tox
-.pytest_cache
-nosetests.xml
-coverage
-doc/_build
-doc/api
-tests/datafile/test_fitter_plot.pdf
-htmlcov
-.coverage.*
-
-# Translations
-*.mo
-
-# Ephemerides and clock stuff
-de405.bsp
-de421.bsp
-de430t.bsp
-de434.bsp
-de436.bsp
-de436t.bsp
-
-earth_latest_high_prec.bpc
-naif0010.tls
-naif0011.tls
-naif0012.tls
-pck00010.tpc
-de-403-masses.tpc
-
-# Pickles of .tim files
-*.tim.pickle.gz
-
-# TEMPO stuff
-*matrix.tmp
-*resid2.tmp
-*tempo.lis
-
-# ipython notebook temps
-.ipynb_checkpoints
-
-.idea
-.cache
-.vscode
-
-# Test output/intermediate files
-# tests/datafile/fake_testzima.tim
-# tests/datafile/fermitest.fits
-# tests/datafile/fermitest.png
-# tests/datafile/photontest.fits
-# tests/datafile/photontest.png
-# tests/datafile/test_parfile_write.par
-# tests/datafile/J0030+0451_chains.png
-# tests/datafile/J0030+0451_post.par
-# tests/datafile/J0030+0451_post.png
-# tests/datafile/J0030+0451_pre.png
-# tests/datafile/J0030+0451_prof_post.txt
-# tests/datafile/J0030+0451_prof_pre.txt
-# tests/datafile/J0030+0451_results.txt
-# tests/datafile/J0030+0451_samples.pickle
-# tests/datafile/J0030+0451_triangle.png
-# tests/datafile/testbary.tim
-# tests/datafile/testtopo.tim
-# tests/datafile/testtopot1.tim
-# fermitest.fits
-# fermitest.png
-# photontest.fits
-# photontest.png
-
-# Allow devs to have a tmp dir that is not tracked
-tmp/*
-
-# Autogenerated rst files
-docs/_build
-docs/_autosummary
-
-# hypothesis testing database
-.hypothesis
-
-# Jupyter notebooks should not be in git; they are built from md files
-docs/examples/*.ipynb
-
-# Actual tests are supposed to be python scripts.
-# These notebooks are used for interactive testing during development.
-# They should be converted to scripts before committing.
-tests/*.ipynb
-
-# Rendered examples are turned into .py files for convenience but it's the notebook that's in git
-docs/examples-rendered/*.py
-
-# VSCode wants to put virtualenvs here
-.env
-
-# pintpublish output
-*.tex
-*.aux
-*.log
-*.pdf
-
-# test output
-tests/datafile/par_*.par
-tests/datafile/fake_toas.tim
-tests/datafile/*.converted.par
-tests/datafile/_test_pintempo.out
