+++ APTchecker.py
